rename file and folder, fix error
