+++ server/controllers/postController.js
@@ -1,4 +1,4 @@
-const postService = require("../services/postController");
+const postService = require("../services/postService");
 
 // get all posts
 exports.getAllPosts = async (req, res) => {
@@ -21,7 +21,7 @@ exports.createPost = async (req, res) => {
 }
 
 // get post by id
-exports.getPost = async (req, res) => {
+exports.getPostById = async (req, res) => {
   try {
     const post = await postService.getPost(req.params.id);
     res.status(200).json(post);
@@ -31,7 +31,7 @@ exports.getPost = async (req, res) => {
 }
 
 // update post by id
-exports.updatePost = async (req, res) => {
+exports.updatePostById = async (req, res) => {
   try {
     const updatedPost = await postService.updatePost(req.params.id, req.body);
     res.status(200).json(updatedPost);
@@ -42,9 +42,9 @@ exports.updatePost = async (req, res) => {
 
 
 // delete post by id
-exports.deletePost = async (req, res) => {
+exports.deletePostById = async (req, res) => {
   try {
-    const deletedPost = await postService.deletePost(id);
+    const deletedPost = await postService.deletePost(req.params.id);
     res.status(200).json({message : "delete success"});
   } catch (err) {
     res.status(500).json({message : err.message})
